requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic-settings==2.6.1
requests==2.32.3
apscheduler==3.10.4
